Send reviews back to WaniKani API.
Added a --dry-run for testing purposes.
This mode does not send anything back to the API.

Added a hard mode that will require the user to input
all the prounonciation of a given kanji.
E.g: なに,なん for 何. The order does not matter.

Changed the entire architecture to handle subject.
Removed Kanji, Vocabulary and Radical class.

Create a class to handle Subject update.