Cache audio during and play it in a thread.
It is useless to download X times the same audio.
We cache it during our first download.

It will change the alternate method a little bit
since caching keeps the same audio gender.

Audio are download at the begining of a
lessons (with extra character).

Audio are played in a different thread.