feat: add windows support
Added mutagen for windows install since playsound has issue playing WaniKani's mp3
Mutagen removes the mp3 tag and that makes it work on windows.

We are not saving temporary files and delete them manually.
It was making permission error on windows.

Updated documentation to include Windows install.

Updated tests when removing cache (since we are keeping the files and delete them manually)

Removed tests that were checking that windows is not implemented.