feat: improve reviews queue management
Reviews are now taken 10 by 10. When completing an item a new item will be added to the queue.

When an item is wrong, the item will go back at the end of the queue after the shuffle.

It avoid the same question to be asked twice in a row. It makes it easier.

Keeping a queue of 10 itwms (max 20 questions or less with radicals), makes it easier to interrupt a session.